requirements.txt
tensorflow>=2.13.0
keras>=3.0.0
numpy>=1.24.0
matplotlib>=3.7.0
pillow>=10.0.0
streamlit>=1.28.0
tensorflow-datasets>=4.9.0
